Warnings

+++ include/effects/matrix/Vector.h
@@ -160,7 +160,7 @@ public:
 
     void rotate(float deg)
     {
-        float theta = deg / 180.0 * M_PI;
+        float theta = deg / 180.0f * (float) M_PI;
         float c = cos(theta);
         float s = sin(theta);
         float tx = x * c - y * s;
@@ -173,7 +173,7 @@ public:
     {
         if (length() == 0)
             return *this;
-        *this *= (1.0 / length());
+        *this *= (1.0f / length());
         return *this;
     }
